server/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
qdrant-client==1.9.0
openai==1.3.5
python-dotenv==1.0.0
requests>=2.31
beautifulsoup4>=4.12
cohere>=4.0
PyYAML>=6.0
tiktoken>=0.7
pytest>=7.0
slowapi>=0.1.5